cookflow_agent/requirements.txt
google-adk==1.4.1
python-dotenv==1.2.0
# google-cloud-logging==3.11.3